requirements.txt
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.2.0
certifi==2026.2.25
charset-normalizer==3.4.5
clip @ git+https://github.com/ultralytics/CLIP.git@88ade288431a46233f1556d1e141901b3ef0a36b
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
easyocr==1.7.2
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
ftfy==6.3.1
hf-xet==1.4.2
huggingface_hub==0.36.2
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
lazy-loader==0.5
MarkupSafe==3.0.3
matplotlib==3.9.4
matplotlib-inline==0.2.1
mistune==3.2.0
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.17.0
nbformat==5.10.4
networkx==3.2.1
ninja==1.13.0
numpy==2.0.2
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
platformdirs==4.4.0
polars==1.36.1
polars-runtime-32==1.36.1
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyclipper==1.3.0.post6
Pygments==2.19.2
pyparsing==3.3.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
rpds-py==0.27.1
safetensors==0.7.0
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.7
six==1.17.0
soupsieve==2.8.3
stack-data==0.6.3
sympy==1.14.0
tifffile==2024.8.30
tinycss2==1.4.0
tokenizers==0.22.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==4.57.6
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.21
ultralytics-thop==2.0.18
urllib3==2.6.3
wcwidth==0.6.0
webencodings==0.5.1
yarg==0.1.9
zipp==3.23.0
